♻️ refactor(home): move about-me config defs to contents
